rename accordion.js to sf-accordion.js
